Backend/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
pymongo==4.6.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
pyotp==2.9.0
django-storages==1.14.2
boto3==1.34.63
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
tzdata==2023.3
